The remove method is an override implementation of the JPA remove method. The implementation that this uses accepts an ID, finds it, and then removes it. No need to do a find before calling remove().
I've also added a simple javascript confirm alert since it's so easy to accidentally click on the Delete link.

+++ src/main/java/oscar/oscarBilling/ca/bc/pageUtil/BillCodesTableWrapper.java
@@ -41,7 +41,7 @@ public class BillCodesTableWrapper
     BillingService bcd = (BillingService) this.getCurrentRowObject();
     String links = "<a href=\"billingEditCode.jsp?codeId=" + bcd.getBillingserviceNo() +
     "&code=" + bcd.getServiceCode()+"&desc=" + bcd.getDescription() + "&value=" + bcd.getValue() + "&whereTo=private\">Edit</a> <br>" +
-    "<a href=\"deletePrivateCode.jsp?code=" + bcd.getBillingserviceNo() + "\">Delete</a>";
+    "<a onclick=\"return confirm('Are you sure you want to delete this code ["+bcd.getServiceCode()+" "+bcd.getDescription()+"]?');\" href=\"deletePrivateCode.jsp?code=" + bcd.getBillingserviceNo() + "\">Delete</a>";
     return links;
   }
